+++ Assets/Scripts/SetCharacterStatus.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class SetCharacterStatus : MonoBehaviour
 {
     [SerializeField, Tooltip("最大体力, 回復, アーマー, 移動速度, 威力, エリア, 速度, 持続時間, 量,　クールダウン, 運気,　成長, 強欲, 呪い, 磁石 初期武器の添え字")]
-    int[] _characterStatus = new int[16];
+    float[] _characterStatus = new float[16];
 
     public void SetStatus()
     {
